extract class PictureUploadController
