+++ src/main/java/com/randy/masterspring/controller/PictureUploadController.java
@@ -1,0 +1,68 @@
+package com.randy.masterspring.controller;
+
+import com.randy.masterspring.config.PictureUploadProperties;
+import org.apache.tomcat.util.http.fileupload.IOUtils;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.core.io.ClassPathResource;
+import org.springframework.core.io.FileSystemResource;
+import org.springframework.core.io.Resource;
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestMethod;
+import org.springframework.web.multipart.MultipartFile;
+import org.springframework.web.servlet.mvc.support.RedirectAttributes;
+
+import javax.servlet.http.HttpServletResponse;
+import java.io.*;
+import java.net.URLConnection;
+
+@Controller
+public class PictureUploadController {
+
+    private static final Resource PICTURES_DIR = new FileSystemResource("./pictures");
+
+    private final Resource picturesDir;
+    private final Resource anonymousPicture;
+
+    @Autowired
+    public PictureUploadController(PictureUploadProperties uploadProperties){
+        picturesDir = uploadProperties.getUploadPath();
+        anonymousPicture = uploadProperties.getAnonymousPicture();
+    }
+
+    @RequestMapping(value = "/uploadedPicture")
+    public void getUploadedPicture(HttpServletResponse response) throws IOException {
+        response.setHeader("Content-Type",URLConnection.guessContentTypeFromName(anonymousPicture.getFilename()));
+        IOUtils.copy(anonymousPicture.getInputStream(),response.getOutputStream());
+    }
+    @RequestMapping("upload")
+    public String uploadPage(){
+        return "profile/uploadPage";
+    }
+
+    @RequestMapping(value = "/upload", method = RequestMethod.POST)
+    public String onUpload(MultipartFile file, RedirectAttributes redirectAttributes) throws IOException{
+
+        if (file.isEmpty() || !isImage(file)) {
+            redirectAttributes.addFlashAttribute("error","Incorrect file. Please upload a picture");
+            return "redirect:/upload";
+        }
+        String fileExtension = getFileExtension(file.getOriginalFilename());
+        File tempFile = File.createTempFile("pic",fileExtension,picturesDir.getFile());
+
+        try (InputStream in = file.getInputStream();
+             OutputStream out = new FileOutputStream(tempFile)){
+            IOUtils.copy(in,out);
+        }
+        return "profile/uploadPage";
+    }
+
+
+    private boolean isImage(MultipartFile file){
+        return file.getContentType().startsWith("image");
+    }
+
+    private static String getFileExtension(String name){
+        return name.substring(name.lastIndexOf("."));
+    }
+}
