+++ src/main/java/com/randy/masterspring/controller/ProfilieController.java
@@ -27,37 +27,6 @@ import java.util.Locale;
 @Controller
 public class ProfilieController {
 
-    private static final Resource PICTURES_DIR = new FileSystemResource("./pictures");
-
-    @RequestMapping(value = "/uploadedPicture")
-    public void getUploadedPicture(HttpServletResponse response) throws IOException {
-        ClassPathResource classPathResource = new ClassPathResource("/images/anonymous.png");
-        response.setHeader("Content-Type", URLConnection.guessContentTypeFromName(classPathResource.getFilename()));
-        IOUtils.copy(classPathResource.getInputStream(),response.getOutputStream());
-    }
-    @RequestMapping("upload")
-    public String uploadPage(){
-        return "profile/uploadPage";
-    }
-
-    @RequestMapping(value = "/upload", method = RequestMethod.POST)
-    public String onUpload(MultipartFile file, RedirectAttributes redirectAttributes) throws IOException{
-
-        if (file.isEmpty() || !isImage(file)) {
-            redirectAttributes.addFlashAttribute("error","Incorrect file. Please upload a picture");
-            return "redirect:/upload";
-        }
-        String fileName = file.getOriginalFilename();
-        File tempFile = File.createTempFile("pic",getFileExtension(fileName),PICTURES_DIR.getFile());
-
-        try (InputStream in = file.getInputStream();
-             OutputStream out = new FileOutputStream(tempFile)){
-            IOUtils.copy(in,out);
-        }
-        return "profile/uploadPage";
-    }
-
-
     @RequestMapping("profile")
     public String displayProfile(ProfileForm profileForm){
         return "profile/profilePage";
@@ -90,11 +59,4 @@ public class ProfilieController {
         return "profile/profilePage";
     }
 
-    private boolean isImage(MultipartFile file){
-        return file.getContentType().startsWith("image");
-    }
-
-    private static String getFileExtension(String name){
-        return name.substring(name.lastIndexOf("."));
-    }
 }
